Basic services for fetching Petfinder API data
We will start by consuming data from the Petfinder API.  The 2nd API
will be Rescue Groups dot Org.  However, the RG API is undergoing an
update and is supposed to be ready by mid-2018.  Because this date is fast
approaching and the current API is lacking many modern features, I will move
onwards before writing the RG API sevices.  There is a clear template
established for adding a new API.  It is my hope that this project
can grow by adding new APIs as they arise or are discovered.

It's worth mentioning that I have been consulting the new RG API
as much as possible in order to make sensible decisions about our
own local DB schema, etc....  For the most part, the simpler data
structure of Petfinder is closer to what we need.  For now, we are
using that Petfinder data as a starting point / guide.